+++ packages/stargate/src/index.ts
@@ -1,5 +1,6 @@
 export { Account, accountFromAny, AccountParser } from "./accounts";
 export { AminoConverter, AminoConverters, AminoTypes } from "./aminotypes";
+export { Attribute, Event, fromTendermint34Event } from "./events";
 export { calculateFee, GasPrice } from "./fee";
 export * as logs from "./logs";
 export {