+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
   supporting batch RPC requests ([#1300]).
 - @cosmjs/encoding: Add `lossy` parameter to `fromUtf8` allowing the use of a
   replacement charater instead of throwing.
+- @cosmjs/stargate: Add structured `Events`s to `IndexTx.events`.
 
 ## [0.29.2] - 2022-10-13